revised to handle cases with no ROIs where get_bin_file now returns targets = NaN
git-svn-id: https://beagle.whoi.edu/svn/ifcb/trunk@642 9f5e5bc8-6430-4f94-ad00-aad15e9f9366

Former-commit-id: d6554696156cc0158f7f9ab5dab7eef65f5bcc0b

+++ feature_extraction/blob_extraction/bin_blobs_heidi.m
@@ -16,11 +16,14 @@ end
 % load the zip file
 log(['LOAD ' file]);
 targets = get_bin_file([in_dir  file]);
-nt = length(targets.targetNumber);
+nt = 0; %default for case with no targets
+if isstruct(targets),
+    nt = length(targets.targetNumber);  
+    png_dir = [out_dir filesep regexprep(file,'.zip','')];
+    mkdir(png_dir);
+end;
 log(['PROCESSING ' num2str(nt) ' target(s) from ' file]);
 
-png_dir = [out_dir filesep regexprep(file,'.zip','')];
-mkdir(png_dir);
 
 png_paths = {};
 % for each target
@@ -38,13 +41,14 @@ for i = 1:nt,
     png_paths = [png_paths; {png_path}];
 end
 
-zip(archive, png_paths, png_dir);
-
-log(['SAVING ' archive]);
-
-rmdir(png_dir,'s');
-
-log(['DONE ' file]);
+if isstruct(targets),
+    zip(archive, png_paths, png_dir);
+    log(['SAVING ' archive]);
+    rmdir(png_dir,'s');
+    log(['DONE ' file]);
+else
+    log(['NO targets ' file]);
+end;
 
 end
